moved matcher tests into own package, split login and public endpoint tests apart from main test file
