+++ src/test/java/com/example/demo/Matcher/OrderServiceTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.Matcher;
 
 import com.example.demo.matcher.models.OrderAction;
 import com.example.demo.matcher.models.OrderbookItem;
